style: move world country helpers out of extensions dir
