+++ packages/sealed_countries/lib/sealed_countries.dart
@@ -14,11 +14,11 @@ export "src/helpers/extensions/country_submodels/gini_extension.dart";
 export "src/helpers/extensions/country_submodels/idd_extension.dart";
 export "src/helpers/extensions/country_submodels/postal_code_extension.dart";
 export "src/helpers/extensions/country_submodels/weekday_extension.dart";
-export "src/helpers/extensions/world_country/world_country_date_time.dart";
-export "src/helpers/extensions/world_country/world_country_geo.dart";
-export "src/helpers/extensions/world_country/world_country_getters.dart";
-export "src/helpers/extensions/world_country/world_country_names.dart";
-export "src/helpers/extensions/world_country/world_country_web_resources.dart";
+export "src/helpers/world_country/world_country_date_time.dart";
+export "src/helpers/world_country/world_country_geo.dart";
+export "src/helpers/world_country/world_country_getters.dart";
+export "src/helpers/world_country/world_country_names.dart";
+export "src/helpers/world_country/world_country_web_resources.dart";
 export "src/model/country/country.dart";
 export "src/model/country/submodels/capital.dart";
 export "src/model/country/submodels/capital_info.dart";
