+++ packages/sealed_countries/lib/src/helpers/world_country/world_country_getters.dart
@@ -1,4 +1,4 @@
-import "../../../model/country/country.dart";
+import "../../model/country/country.dart";
 
 /// A class extension that adds utility methods to the [WorldCountry] class.
 extension WorldCountryGetters on WorldCountry {
