+++ packages/sealed_countries/lib/src/helpers/world_country/world_country_web_resources.dart
@@ -1,4 +1,4 @@
-import "../../../model/country/country.dart";
+import "../../model/country/country.dart";
 
 /// Provides extension methods for accessing web resources related to a
 /// `WorldCountry` object.
