+++ packages/sealed_countries/test/src/helpers/world_country/world_country_names_test.dart
@@ -1,4 +1,4 @@
-import "package:sealed_countries/src/helpers/extensions/world_country/world_country_names.dart";
+import "package:sealed_countries/src/helpers/world_country/world_country_names.dart";
 import "package:sealed_countries/src/model/country/country.dart";
 import "package:sealed_currencies/sealed_currencies.dart";
 import "package:test/test.dart";
